+++ src/inputs/datepicker/datepicker.component.ts
@@ -1,7 +1,7 @@
-import {Component, Input} from '@angular/core'
-
 import {BaseInputComponent} from '../base/baseInput.component'
+import {Component, Input} from '@angular/core'
 import {DatepickerFieldDataInterface} from './datepicker.interfaces'
+import {MatDatepicker} from '@angular/material'
 
 
 @Component({
@@ -13,12 +13,18 @@ export class DatepickerComponent extends BaseInputComponent {
 	@Input()
 	fieldData: DatepickerFieldDataInterface
 
-
 	constructor() {
 		super()
 	}
 
 	ngOnInit(): void {
 		super.ngOnInit()
+	}
+
+	openCalendar(picker: MatDatepicker<Date>): void {
+		if (this.fieldData.readOnly) {
+			return
+		}
+		picker.open()
 	}
 }